criado crud da credencial do codigo de barras, ajustado cadastrar usuario, excluir endereco, iniciado hateos

+++ automanager/src/main/java/com/autobots/automanager/entidades/CredencialCodigoBarra.java
@@ -11,13 +11,13 @@ import lombok.EqualsAndHashCode;
 @Entity
 public class CredencialCodigoBarra extends Credencial {
 	@Column(nullable = false, unique = true)
-	private long codigo;
+	private double codigo;
 
-	public long getCodigo() {
+	public double getCodigo() {
 		return codigo;
 	}
 
-	public void setCodigo(long codigo) {
+	public void setCodigo(double codigo) {
 		this.codigo = codigo;
 	}
 }
